backend/requirements.txt
fastapi==0.104.1
uvicorn
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
cors==1.0.1 
